Feature/tp01

+++ services/user-svc/src/main/java/edu/utn/frsf/isi/dan/user/dao/CuentaBancariaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CuentaBancariaRepository extends JpaRepository<CuentaBancaria, Long> {
+public interface CuentaBancariaRepository extends JpaRepository<CuentaBancaria, Integer> {
 }
